# functions/Swagger/openapi.yaml
openapi: 3.0.1
info:
  version: "1.00"
  title: 25 Days of Serverless
  description: Microsoft's 25 Days of Serverless challenge but with Google Cloud Functions for Firebase
  contact:
    name: Nipuna Weerasekara
    email: w.nipuna@gmail.com
    url: "https://niweera.gq/"
servers:
  - url: "https://us-central1-fsuptutorial.cloudfunctions.net/api/"
    description: Production server
tags:
  - name: READ operations
paths:
  /:
    get:
      tags:
        - READ operations
      description: Get all endpoints
      operationId: rootEndPoint
      parameters: []
      responses:
        "200":
          description: All endpoints are returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootEndPoints"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /one:
    get:
      tags:
        - READ operations
      description: Get the Challenge One
      operationId: challengeOne
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /four:
    get:
      tags:
        - READ operations
      description: Get the Challenge Four
      operationId: challengeFour
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /four/add:
    post:
      tags:
        - CREATE operations
      description: Add new Dish to the Database
      operationId: challengeFour_create
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /four/list:
    get:
      tags:
        - READ operations
      description: List all dishes from the Database
      operationId: challengeFour_read
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /four/change:
    patch:
      tags:
        - UPDATE operations
      description: List all dishes from the Database
      operationId: challengeFour_update
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /four/remove:
    delete:
      tags:
        - DELETE operations
      description: List all dishes from the Database
      operationId: challengeFour_delete
      parameters: []
      responses:
        "200":
          description: Return the Challenge One Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/challengeOne"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
  /api-docs/:
    get:
      tags:
        - READ operations
      description: API Documentation
      operationId: apiDocumentation
      parameters: []
      responses:
        "200":
          description: API Documentation is returned
          content:
            text/html:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Internal Server Error
                code: internal
components:
  schemas:
    RootEndPoints:
      type: object
      properties:
        rootEndPoint:
          type: string
          example: /
        challengeOne:
          type: string
          example: /one
        apiDocumentation:
          type: string
          example: /api-docs
      description: Root Endpoints of Bug Zero API
      example:
        rootEndPoint: /
        challengeOne: /one
        apiDocumentation: /api-docs
    challengeOne:
      type: object
      properties:
        letter:
          type: string
          example: "ש"
      description: The letter
      example:
        letter: "ש"
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        name:
          type: string
        stack:
          type: string
      example:
        message: Internal Server Error
        code: internal
        name: Internal Server Error
        stack: stack